Refactor into more specialized components
