+++ haufe-fe/src/components/error/HealthCheckError.component.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export function HealthCheckErrorComponent() {
+    return (
+        <><h1>NOT ALLOWED</h1><h3>Backend is not healthy!</h3></>
+    )
+}
