+++ haufe-fe/src/components/home/Home.component.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+export function HomeComponent() {
+    return (
+            <>
+                <h1>Haufe Group - FE app</h1>
+                <p className="lead">Welcome to this main page! :)</p>
+                <p className="lead">
+                    <a href="#" className="btn btn-lg btn-secondary fw-bold border-white bg-white">Learn more</a>
+                </p>
+            </>
+    )
+}
