+++ haufe-fe/src/components/sections/app.main.tsx
@@ -1,14 +1,48 @@
 import React from "react";
+import {HTTP} from "../../common/http";
+import {StatusMessage} from "../domain/status-message.model";
+import {HomeComponent} from "../home/Home.component";
+import {HealthCheckErrorComponent} from "../error/HealthCheckError.component";
 
-export class AppMain extends React.Component<{}, {}> {
+interface StateModel {
+    isLoading: boolean | undefined,
+    isBackendHealthy: boolean | undefined
+}
+
+export class AppMain extends React.Component<{}, StateModel> {
+
+    constructor(props: any) {
+        super(props);
+        this.state = {
+            isLoading: undefined,
+            isBackendHealthy: undefined
+        }
+    }
+
+    componentDidMount() {
+        this.setState({isLoading: true});
+
+        HTTP.get('/health').then(
+            (response: StatusMessage) => {
+                this.setState((state, _) => ({
+                    isBackendHealthy: response.status_code >= 200 && response.status_code < 300
+                }));
+            },
+            (reason: any) => {
+                console.error('Backend is not healthy:', reason);
+                this.setState({
+                    isBackendHealthy: false
+                })
+            }
+        ).finally(() => this.setState({isLoading: false}))
+    }
+
     render() {
+        const isBackendHealthy = this.state.isBackendHealthy;
+
         return (
             <main className="px-3">
-                <h1>Haufe Group - FE app</h1>
-                <p className="lead">Welcome to this main page! :)</p>
-                <p className="lead">
-                    <a href="#" className="btn btn-lg btn-secondary fw-bold border-white bg-white">Learn more</a>
-                </p>
+                { isBackendHealthy ? <HomeComponent/> : <HealthCheckErrorComponent/>}
             </main>
         )
     }
